fix: keep chat section color styling also on area view + rm unused dialogue-turn properties
