+++ interfaces/IBF-dashboard/src/app/components/chat/chat.component.html
@@ -1,5 +1,5 @@
 <ion-content scrollY="true">
-  <app-dialogue-turn [isWarn]="isWarn" [isOpeningBubble]="true">
+  <app-dialogue-turn [isLate]="isLate">
     <p style="padding-right: 20px">
       <span
         [innerHTML]="
@@ -11,7 +11,7 @@
                 }
         "
       ></span>
-      @if (isWarn) {
+      @if (isLate) {
         <app-tooltip
           [value]="
             'chat-component.common.warn-label.tooltip'
@@ -56,6 +56,7 @@
       [adminAreaLabelPlural]="adminAreaLabelPlural"
       [mainExposureIndicatorLabel]="mainExposureIndicatorLabel"
       [mainExposureIndicatorNumberFormat]="mainExposureIndicatorNumberFormat"
+      [cardColors]="getCardColors(event)"
     ></app-event-speech-bubble>
   } @else {
     <!-- Active event -->
@@ -74,13 +75,17 @@
             mainExposureIndicatorNumberFormat
           "
           [forecastSource]="countryDisasterSettings?.forecastSource"
+          [cardColors]="getCardColors(event)"
         ></app-event-speech-bubble>
       }
     }
   }
 
   @for (area of filteredAreas; track area.name; let isFirst = $first) {
-    <app-dialogue-turn [isTriggered]="eventState?.event?.forecastTrigger">
+    <app-dialogue-turn
+      [borderColor]="getCardColors(eventState?.event).borderColor"
+      [isSelected]="true"
+    >
       <ion-row>
         <ion-col
           class="flex flex-row gap-2"
@@ -88,47 +93,30 @@
           size-md="12"
           size-sm="8"
         >
-          <ion-img
-            src="assets/icons/Alert_Chat_Black.svg"
-            class="chat-icon"
-          ></ion-img>
-          <div>
-            <ion-label
-              [innerHTML]="
-                'chat-component.' +
-                  disasterTypeName +
-                  '.active-event.place-name'
-                  | translate
-                    : {
-                        adminAreaLabel,
-                        placeName: area.name,
-                        parentName: getAreaParentString(area),
-                      }
-              "
-            ></ion-label
-            ><br />
-            @if (area.mainExposureValue) {
-              <ion-note size="small">
-                <span
-                  [innerHTML]="
-                    'chat-component.' +
-                      disasterTypeName +
-                      '.active-event.exposed' | translate
-                  "
-                ></span>
-                <strong>
-                  {{
-                    area.mainExposureValue
-                      | compact: mainExposureIndicatorNumberFormat
-                  }}
-                </strong>
-              </ion-note>
-            } @else {
-              <ion-note size="small">
-                <span>{{ 'chat-component.common.no-info' | translate }}</span>
-              </ion-note>
-            }
-            <br />
+          <div class="event-header">
+            <span class="event-header--icon-title">
+              <ion-icon
+                [style.color]="getCardColors(eventState?.event).iconColor"
+                class="alert-icon"
+                src="/assets/icons/Alert_Title.svg"
+                slot="start"
+              ></ion-icon>
+              <span
+                [style.color]="getCardColors(eventState?.event).headerTextColor"
+                [innerHTML]="
+                  'chat-component.' +
+                    disasterTypeName +
+                    '.active-event.place-name'
+                    | translate
+                      : {
+                          adminAreaLabel,
+                          placeName: area.name,
+                          parentName: getAreaParentString(area),
+                        }
+                "
+              >
+              </span>
+            </span>
           </div>
         </ion-col>
         <ion-col size-lg="4" size-md="12" size-sm="4">
@@ -147,6 +135,26 @@
           }
         </ion-col>
       </ion-row>
+      @if (area.mainExposureValue) {
+        <ion-note size="small">
+          <span
+            [innerHTML]="
+              'chat-component.' + disasterTypeName + '.active-event.exposed'
+                | translate
+            "
+          ></span>
+          <strong>
+            {{
+              area.mainExposureValue
+                | compact: mainExposureIndicatorNumberFormat
+            }}
+          </strong>
+        </ion-note>
+      } @else {
+        <ion-note size="small">
+          <span>{{ 'chat-component.common.no-info' | translate }}</span>
+        </ion-note>
+      }
       @if (
         eventState?.event?.forecastTrigger &&
         countryDisasterSettings.enableEarlyActions &&
