+++ interfaces/IBF-dashboard/src/app/components/date-button/date-button.component.html
@@ -15,7 +15,7 @@
 </span>
 @if (forecastAlert) {
   <ion-icon
-    class="warning-icon"
+    class="alert-icon"
     [src]="
       forecastTrigger
         ? '/assets/icons/Alert_Title_Purple.svg'
