+++ interfaces/IBF-dashboard/src/app/components/date-button/date-button.component.scss
@@ -13,7 +13,7 @@
   float: left;
 }
 
-.warning-icon {
+.alert-icon {
   vertical-align: middle;
   margin-left: 5px;
 }
