+++ interfaces/IBF-dashboard/src/app/components/dialogue-turn/dialogue-turn.component.scss
@@ -29,7 +29,7 @@
     padding-bottom: 8px !important;
   }
 }
-.dialogue-turn.is-warn {
+.dialogue-turn.is-late {
   .dialogue-turn--content {
     ion-card-content {
       color: var(--ion-color-danger);
@@ -41,16 +41,7 @@
   }
 }
 
-.dialogue-turn.is-pending {
-  opacity: 0;
-
-  .dialogue-turn--content {
-    opacity: 0;
-    transform: scale(0.2);
-  }
-}
-
-.dialogue-turn.is-spoken {
+.dialogue-turn {
   .dialogue-turn--content {
     animation-name: speak;
     animation-duration: 0.3s;
