+++ interfaces/IBF-dashboard/src/app/components/event-speech-bubble/event-speech-bubble.component.scss
@@ -30,7 +30,7 @@ ion-checkbox {
   }
 }
 
-.warning-icon {
+.alert-icon {
   float: left;
   margin-right: 5px;
   margin-top: 3px;
