+++ interfaces/IBF-dashboard/src/app/components/event-speech-bubble/event-speech-bubble.component.ts
@@ -3,6 +3,7 @@ import { PopoverController } from '@ionic/angular';
 import { TranslateService } from '@ngx-translate/core';
 import { Subscription } from 'rxjs';
 import { AuthService } from 'src/app/auth/auth.service';
+import { CardColors } from 'src/app/components/chat/chat.component';
 import { SetTriggerPopoverComponent } from 'src/app/components/set-trigger-popover/set-trigger-popover.component';
 import { DisasterType, ForecastSource } from 'src/app/models/country.model';
 import { PlaceCode } from 'src/app/models/place-code.model';
@@ -41,6 +42,8 @@ export class EventSpeechBubbleComponent implements AfterViewChecked, OnDestroy {
   public mainExposureIndicatorLabel: string;
   @Input()
   public mainExposureIndicatorNumberFormat: NumberFormat;
+  @Input()
+  public cardColors: CardColors;
 
   public typhoonLandfallText: string;
   public displayName: string;
@@ -159,55 +162,6 @@ export class EventSpeechBubbleComponent implements AfterViewChecked, OnDestroy {
     }
 
     return false;
-  }
-
-  public getCardColors(): {
-    iconColor: string;
-    headerTextColor: string;
-    borderColor: string;
-  } {
-    const defaultColors = {
-      iconColor: 'var(--ion-color-ibf-black)',
-      headerTextColor: 'var(--ion-color-ibf-black)',
-      borderColor: null,
-    };
-
-    if (!this.event) {
-      return defaultColors;
-    }
-
-    if (!this.event.disasterSpecificProperties) {
-      if (!this.event.forecastTrigger) {
-        return defaultColors;
-      }
-
-      return {
-        iconColor: 'var(--ion-color-fiveten-red-500)',
-        headerTextColor: 'var(--ion-color-fiveten-red-500)',
-        borderColor: 'var(--ion-color-fiveten-red-500)',
-      };
-    }
-
-    if (!this.event.disasterSpecificProperties.eapAlertClass) {
-      if (!this.event.forecastTrigger) {
-        return defaultColors;
-      }
-
-      return {
-        iconColor: 'var(--ion-color-fiveten-red-500)',
-        headerTextColor: 'var(--ion-color-fiveten-red-500)',
-        borderColor: 'var(--ion-color-fiveten-red-500)',
-      };
-    }
-
-    return {
-      iconColor: `var(--ion-color-${this.event.disasterSpecificProperties.eapAlertClass.color})`,
-      headerTextColor: `var(--ion-color-${
-        this.event.disasterSpecificProperties.eapAlertClass.textColor ||
-        this.event.disasterSpecificProperties.eapAlertClass.color
-      })`,
-      borderColor: `var(--ion-color-${this.event.disasterSpecificProperties.eapAlertClass.color})`,
-    };
   }
 
   public async openSetTriggerPopover(): Promise<void> {
